+++ src/client/OrganizationAccountClient.ts
@@ -38,14 +38,13 @@ import {
 import {
   ACCOUNT_QUERY_STRING,
   REPOS_QUERY_STRING,
-  TEAM_MEMBERS_QUERY_STRING,
-  TEAM_REPOS_QUERY_STRING,
+  SINGLE_TEAM_MEMBERS_QUERY_STRING,
   ISSUES_QUERY_STRING,
   PULL_REQUESTS_QUERY_STRING,
   TEAMS_QUERY_STRING,
   USERS_QUERY_STRING,
-  COLLABORATORS_QUERY_STRING,
   SINGLE_REPO_COLLABORATORS_QUERY_STRING,
+  SINGLE_TEAM_REPOS_QUERY_STRING,
 } from './GraphQLClient/queries';
 
 export default class OrganizationAccountClient {
@@ -162,15 +161,20 @@ export default class OrganizationAccountClient {
     return response;
   }
 
-  async getTeamMembers(): Promise<OrgTeamMemberQueryResponse[]> {
+  async getTeamMembers(
+    teamSlug: string,
+  ): Promise<OrgTeamMemberQueryResponse[]> {
     let response: OrgTeamMemberQueryResponse[] = [];
 
     await this.queryGraphQL('team members', async () => {
       const { members, rateLimitConsumed } = await this.v4.fetchFromSingle(
-        TEAM_MEMBERS_QUERY_STRING,
+        SINGLE_TEAM_MEMBERS_QUERY_STRING,
         GithubResource.Organization,
         [GithubResource.TeamMembers],
-        { login: this.login },
+        {
+          login: this.login,
+          slug: teamSlug,
+        },
       );
 
       if (members) {
@@ -208,44 +212,26 @@ export default class OrganizationAccountClient {
     }
   }
 
-  async getTeamRepositories(): Promise<OrgTeamRepoQueryResponse[]> {
+  async getTeamRepositories(
+    teamSlug: string,
+  ): Promise<OrgTeamRepoQueryResponse[]> {
     let response: OrgTeamRepoQueryResponse[] = [];
-
     await this.queryGraphQL('team repositories', async () => {
       const { teamRepositories, rateLimitConsumed } =
         await this.v4.fetchFromSingle(
-          TEAM_REPOS_QUERY_STRING,
+          SINGLE_TEAM_REPOS_QUERY_STRING,
           GithubResource.Organization,
           [GithubResource.TeamRepositories],
-          { login: this.login },
+          {
+            login: this.login,
+            slug: teamSlug,
+          },
         );
 
       if (teamRepositories) {
         response = response.concat(
           teamRepositories as OrgTeamRepoQueryResponse[],
         );
-      }
-
-      return rateLimitConsumed;
-    });
-
-    return response;
-  }
-
-  async getCollaborators(): Promise<Collaborator[]> {
-    let response: Collaborator[] = [];
-
-    await this.queryGraphQL('collaborators', async () => {
-      const { collaborators, rateLimitConsumed } =
-        await this.v4.fetchFromSingle(
-          COLLABORATORS_QUERY_STRING,
-          GithubResource.Organization,
-          [GithubResource.Collaborators],
-          { login: this.login },
-        );
-
-      if (collaborators) {
-        response = response.concat(collaborators as Collaborator[]);
       }
 
       return rateLimitConsumed;